Add files via upload
The primary objective of this small experiment is to delve into the relationship between the string length, frequency, and period of a simple pendulum. I will determine how a pendulum's frequency (the number of oscillations per unit time) and period (the time taken for one complete oscillation) depend on the string length (L). Using Python and plotting different graphs and animating the motion of the pendulum, I aim to provide a comprehensive visual representation of this relationship. I used ''numpy'' and ''matplotlib.pyplot'' and ''time'' libraries. Although I wrote some codes for the animation part, too; however, the animated part isn't visible